added cli install
